+++ spec/scanner/errors_spec.rb
@@ -64,4 +64,9 @@ RSpec.describe(Regexp::Scanner) do
   include_examples 'scan error', RS::InvalidBackrefError, 'empty refcall', '\g<>'
   include_examples 'scan error', RS::InvalidBackrefError, 'empty refcall', '\g\'\''
   include_examples 'scan error', RS::UnknownUnicodePropertyError, 'unknown property', '\p{foobar}'
+  include_examples 'scan error', RS::UnknownPosixClassError, 'unknown POSIX class [::]', '[[::]]'
+  include_examples 'scan error', RS::UnknownPosixClassError, 'unknown POSIX class [:^:]', '[[:^:]]'
+  include_examples 'scan error', RS::UnknownPosixClassError, 'unknown POSIX class [:x:]', '[[:x:]]'
+  include_examples 'scan error', RS::UnknownPosixClassError, 'unknown POSIX class', '[[:^x:]]'
+  include_examples 'scan error', RS::UnknownPosixClassError, 'unknown POSIX class', '[[:WORD:]]'
 end